CODE 2 / 08-09-2022
Q1. Find PIN: You are given a Math Game. You are provided with 4 numbers: input1, input2, input3 and input4 and expected to find a secret PIN. 
All three numbers input1, input2, and input3 are four-digit numbers within the range >=1000 and <=9999 and input4 is a positive integer. 
PIN= (MAX digit of input1 * MIN digit of input1) + (MAX digit of input2 * MIN digit of input2) + (MAX digit of input3 * MIN digit of input3)-input4
Example- if input1=3521, input2= 2452, input3=1352 and input4=38 then PIN= ((5*1) + (5*2) + (5*1))-38=-18
Assume that the four numbers are passed to the given function and return the PIN. 
Constraints:
1.	If any of the input is not within the specified range return -999 and display “incorrect data”
2.	No number can be negative