CODE 4 / 09-09-2022
Find the sum of all prime numbers except the largest prime: Madhav has been given a task of finding the sum of the prime numbers in a given array, except the largest prime in the array. Madhav approaches you for help to do this by writing a program. 
Example: if input1[]={10,41,18,50,43,31,29,25,59,96,67} represents the array and input2 is 11 represents the number of elements in the array, then the expected output is 203, which is the sum of all prime numbers in the array except the largest which is 67. 
Explanation: Prime numbers: 41, 43, 31, 29, 59, and 67. Thus output= 41+43+31+29+59=203.
Special Condition:
If the array does not contain any prime numbers, then the output should be the sum of all numbers except the largest number. Example:  If input1[]={10,20,30,40} and input2=4, then the output=10+20+30=60