CODE 6 / 10-09-2022
Calculate the sum of non-prime index values in an array: Given an array of N elements, you are expected to find the sum of the values that are present in non-prime indexes of the array. Note that the array index starts from 0 and so on. 
Example1: if input1[]= {10,20,30,40,50,60,70,80,90,100} and input2=10, then values at non-prime indexes are 10, 20, 50, 70, 90, 100 and sum=340.
Example2: if input1[]={-1,-2,-3,3,4,-7} and input2=6 then the values at non-prime indexes are -1, -2, 4 and their sum=1.
Example3: in input1[]={-4,-2} and input2=2, then the values at non-prime indexes are -4,-2 and their sum=-6