CODE 8 / 11-09-2022
Q5. Find Password: There are five numbers available with a kid. He wants to create a password with these five numbers. These numbers are either stable or unstable. A number is said to be stable if each of the distinct digit occur the same number of times, i.e the frequency of each distinct digit in the number is same. For example: 2277, 4004, 11, 23, 583835, 1010 etc are example of stable numbers. Similarly, a number is unstable if the frequency of each digit in the number is NOT the same. For example: 221, 4314, 101, 233, 58135, 101 etc are example of unstable numbers. 
The password can be created as follows: 
Password= Maximum of all stable numbers-Minimum of all stable numbers 
Assume that the five numbers are represented by input1, input2, input3, input4 and input5. Complete the function and return the password. 
For example: if input1=12, input2=1313, input3=122, input4=678, and input5= 898 
Stable numbers are: 12, 1313, and 678   Unstable numbers are: 122, and 898
So, the password should be= 1313-12=1301