Bracket Match Code / 16-09-2022
A string of brackets is considered correctly matched if every opening bracket in the string can be paired up with a later closing bracket, and vice versa. For instance, “(())()” is correctly matched, whereas “)(“ and “((” aren’t. For instance, “((” could become correctly matched by adding two closing brackets at the end, so you’d return 2.

Given a string that consists of brackets, write a function bracket_match that takes a bracket string as an input and returns the minimum number of brackets you’d need to add to the input in order to make it correctly matched.

Test Case 1 -

Input -  “(()”
Output - 1

Test Case 2 -

Input - “(())”
Output - 0

Test Case - 

Input - “())(”
Output - 2