Create Q6.java
Rajat has an array A[] of N integers whose all elements are 0 initially. He wants you to perform Q
operations on it.
Types of operations:
1 X: Update the value of A[X] to 2 ∗ A[X] + 1.
2 X: Update the value of A[X] to f(A[X]/2), where f is Greatest Integer Function:- The greatest integer
function is a function that gives the largest integer which is less than or equal to x. This function is denoted
by ⌊ x⌋ . We will round off the given number to the nearest integer that is less than or equal to the number
itself.
3 X Y: Take all the A[i] such that X<= i <= Y and sum them.
Input:
1) First line will contain two integers N and Q.
2) Next Q lines will contain the operations as described above.
For example if Q lines contains the below input:-
1 2 -> Here 1 means type of operation to do and 2 here means the index of the array
3 1 2 -> Here 3 means type of operation to do and 1 2 here indicates the range of index of the array on
which operation will be done.
Output:
For each type-3 operation, print the sum obtained from the type-3 operation.
Constraints
1) 1<=N, 3<=Q<=500000
2) 1<=X<=Y<=N
3) There will be at least 1 type-1, type-2 and type-3 operation in the each test case.
Scoring
Score is the percentage of correctly solved sets.
Sample Input
5 3
1 1
1 2
3 1 3
Sample Output
2